+++ components/mini-explore/map/index.jsx
@@ -11,7 +11,7 @@ import { useFetchDatasets } from 'hooks/dataset/fetch-datasets';
 import { useGeostore } from 'hooks/geostore';
 
 // constants
-import { USER_AREA_LAYER_TEMPLATES } from 'components/map/constants';
+import { USER_AREA_LAYER_TEMPLATES, BASEMAP_LABEL_DICTIONARY } from 'components/map/constants';
 
 // utils
 import { getUserAreaLayer } from 'components/map/utils';
@@ -249,8 +249,9 @@ export default function MiniExploreMapContainer({
     (_basemap) => {
       const { id } = _basemap;
       dispatch(setBasemap(id));
+      if (labelsId !== 'none') dispatch(setLabels(BASEMAP_LABEL_DICTIONARY[id]));
     },
-    [dispatch],
+    [dispatch, labelsId],
   );
 
   const handleResetView = useCallback(() => {